requirements.txt
flask
flask-login
flask-wtf
flask-limiter
psycopg2-binary
firebase-admin
supabase
python-dotenv
pytest